+++ CHANGES.rst
@@ -1,3 +1,31 @@
+Changes in 1.0.4 (2020-08-07)
+=================================================
+
+Features:
+ * 
+
+Improvements:
+ * Upgrade MatrixKit version ([v0.12.11](https://github.com/matrix-org/matrix-ios-kit/releases/tag/v0.12.11)).
+ * 
+
+Bugfix:
+ * 
+
+API Change:
+ * 
+
+Translations:
+ * 
+
+Others:
+ * 
+
+Build:
+ * 
+
+Test:
+ * 
+
 Changes in 1.0.3 (2020-08-05)
 ===============================================
 
@@ -12,10 +40,12 @@ Improvements:
  * AuthenticationVC: Make custom server options and register button configurable.
  * Xcconfig: Add product bundle identifiers for each target.
  * BuildSettings: Namespace some settings.
+ * BuildSettings: Reuse base bundle identifier for various settings.
 
 Bug fix:
  * Rebranding: Remove Riot from app name (#3497).
  * AuthenticationViewController: Fix custom homeserver textfield scroll issue (#3467).
+ * Rebranding: Update provisioning universal link domain (#3483).
 
 Changes in 1.0.2 (2020-07-28)
 ===============================================